Split settings into subroutes with shared shell
- Add dedicated style subroute and shared settings navigation shell
- Consolidate settings reset state and restore-defaults handling
- Keep the sidebar active on nested settings pages
